requirements.txt
langchain
langchain-google-genai
langchain-chroma
langgraph
chromadb
pydantic
python-dotenv
tiktoken
